add mips
