+++ src/tests/mips.rs
@@ -1,0 +1,8 @@
+use crate::isa_test;
+use crate::Decoder;
+use std::fs::read_to_string;
+
+#[test]
+fn test_mips() {
+    isa_test!("../../examples/mips.toml", "test_data/mips/mips.test");
+}
