+++ src/tests/mod.rs
@@ -1,1 +1,3 @@
+mod mips;
 mod rv32;
+mod util;
